move api to service
